database split - moved db_update
