+++ arm/ui/database/database.py
@@ -2,19 +2,22 @@
 ARM route blueprint for database pages
 Covers
 - database [GET]
+- dbupdate [POST]
+- import_movies [JSON]
 """
 
 import os
 import json
 import re
 from flask_login import LoginManager, login_required  # noqa: F401
-from flask import render_template, request, Blueprint
+from flask import render_template, request, Blueprint, flash, redirect
 
 import arm.ui.utils as ui_utils
 from arm.ui import app, db, constants
 from arm.models import models as models
 import arm.config.config as cfg
 from arm.ui.metadata import get_omdb_poster
+from arm.ui.forms import DBUpdate
 
 route_database = Blueprint('route_database', __name__,
                         template_folder='templates',
@@ -47,6 +50,31 @@ def view_database():
     return render_template('databaseview.html',
                            jobs=jobs.items, pages=jobs,
                            date_format=cfg.arm_config['DATE_FORMAT'])
+
+
+@route_database.route('/dbupdate', methods=['POST'])
+def update_database():
+    """
+    Update the ARM database when changes are made or the arm db file is missing
+    """
+    form = DBUpdate(request.form)
+    if request.method == 'POST' and form.validate():
+        if form.dbfix.data == "migrate":
+            app.logger.debug("User requested - Database migration")
+            ui_utils.arm_db_migrate()
+            flash("ARM database migration successful!", "success")
+        elif form.dbfix.data == "new":
+            app.logger.debug("User requested - New database")
+            flash("ARM database setup successful!", "success")
+        else:
+            # No method defined
+            app.logger.debug(f"No update method defined from DB Update - {form.dbfix.data}")
+            flash("Error no update method specified, report this as a bug.", "error")
+
+        return redirect('/index')
+    else:
+        # Catch for GET requests of the page, redirect to index
+        return redirect('/index')
 
 
 @route_database.route('/import_movies')
